Describe how to configure a remote host to relay via MIAB
Since the conclusion was that #212 was the wrong approach for relaying, I've
recently dropped that patch from my deployment in order to keep as close to
vanilla upstream as possible.

So I'm now doing SASL based relaying, and thought it would be nice to document
how to do so somewhere.