fix LTDs using multiple digests for DS record
Some LTDs like .be use both digest functions 1 and 2 resulting in 2 DS records. This eventuality was not taken into account by the script resulting in a false warning.

This quick fix does not address the problem caused if multiple DNSSEC keys are configured.
This should be addressed as it could be used to migrate users smoothly to a new DNSSEC algorithm, or to rotate the keys as suggested by the DNSSEC specifications.