Add TLSA to the list of allowed custom records
This commit allows custom TLSA records to be added to the
managed domains. Since the criteria for the primary domain TLSA
record is 3 1 1 and the default key for all certificates is the
same one, the same TLSA record generated for the primary domain
will work for all managed domains.

Note that this commit does not modify the custom DNS configuration
page. It only permits TLSA records to be added. Currently the only
way to do that is by manually editing the custom.yaml file in
/home/user-data/dns.