Add TLSA records for locally-hosted domains
This commit adds TLSA records for all locally-hosted (sub)domains
(those with A or AAAA records that match the local box).

TLSA records are computed using the build_tlsa_record() method
based on the default certificate (or key). Since the default
key is the same for all domains hosted on the box, and the TLSA
criteria used are 3 1 1, the TLSA record will be the same for all
domains.

At this point, there is no opt-out mechanism. I'm not sure if this
is necessary, nor what would be the best way to do it if it is
necessary.