Switch PHP's default process manager
Changing from dynamic to ondemand lowers the number of waiting processes from the default of 3 to 1. PHP automatically scales up the number of processes when required, but kills them after 10 seconds (default). I've seen PHP's memory consumption drop significantly. This should be very helpful for low-end VPS servers.

Based on this article: https://ma.ttias.be/a-better-way-to-run-php-fpm/ and tested for a couple of months on a VPS with 1 GB. No noticeable negative loading times, but about 15% less RAM usage.