Dynamic DNS: Allow overriding http method in request parameter
I own a DSL router that allows me to specify a custom url for
dynamic dns updates, but which can only do GET requests.

In order to be able to update addresses on a mailinabox server, I added
the option to emulate the required request method bu specifying it in a
request parameter.