sieve-spam.txt do not stop evaluation
Hi 

I want to be able to discard messages with "X-Spam-Status" = ^YES

To do that I create a sieve in roundcube mail. 

But since the evaluation of Spam e-mails is stopped at a very early step of the processing the roundcube dovecot filter does not apply.

To make that possible the `stop;`command is removed.