Allow to backup on any S3-compatible service
Starting from the versions subsequent to the v57a, the default backend
of duplicity for backups to S3-compatible services is boto3
which requires the indication of the region (--S3-region-name)
for most cases in which the S3 service is not provided by AWS.
Thanks to this intervention, the region of the S3 service
is provided as a parameter to duplicity and is stored in the
/home/user-data/backup/custom.yaml file, like any other backup
configuration parameters.
Moreover, the list of the regions shown to the user in the drop-down
menu in the backup configuration panel is replaced with
a normal text field, since the list included the AWS regions only
and therefore was not usable in all other cases.

See issue #2200.